+++ packages/renderer/src/renderer/screens/Action/DropdownMenus/index.tsx
@@ -1,0 +1,5 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export * from './ExtrinsicDropdownMenu';
+export * from './DropdownMenuDemo';
